+++ src/Events/ModelWasUnliked.php
@@ -11,7 +11,7 @@
 
 namespace Cog\Likeable\Events;
 
-use Cog\Likeable\Contracts\HasLikes as HasLikesContract;
+use Cog\Likeable\Contracts\Likeable as LikeableContract;
 
 /**
  * Class ModelWasUnliked.
@@ -23,7 +23,7 @@ class ModelWasUnliked
     /**
      * The unliked model.
      *
-     * @var \Cog\Likeable\Contracts\HasLikes
+     * @var \Cog\Likeable\Contracts\Likeable
      */
     public $model;
 
@@ -37,13 +37,13 @@ class ModelWasUnliked
     /**
      * Create a new event instance.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @param int $likerId
      * @return void
      */
-    public function __construct(HasLikesContract $model, $likerId)
+    public function __construct(LikeableContract $likeable, $likerId)
     {
-        $this->model = $model;
+        $this->model = $likeable;
         $this->likerId = $likerId;
     }
 }