+++ src/Services/LikeableService.php
@@ -11,14 +11,15 @@
 
 namespace Cog\Likeable\Services;
 
-use Cog\Likeable\Contracts\HasLikes as HasLikesContract;
+use Cog\Likeable\Contracts\Likeable as LikeableContract;
 use Cog\Likeable\Contracts\Like as LikeContract;
 use Cog\Likeable\Contracts\LikeableService as LikeableServiceContract;
 use Cog\Likeable\Contracts\LikeCounter as LikeCounterContract;
 use Cog\Likeable\Enums\LikeType;
 use Cog\Likeable\Exceptions\LikerNotDefinedException;
 use Cog\Likeable\Exceptions\LikeTypeInvalidException;
-use DB;
+use Illuminate\Database\Query\JoinClause;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Database\Eloquent\Builder;
 
 /**
@@ -31,23 +32,23 @@ class LikeableService implements LikeableServiceContract
     /**
      * Add a like to likeable model by user.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @param string $type
      * @param string $userId
      * @return void
      *
      * @throws \Cog\Likeable\Exceptions\LikerNotDefinedException
      */
-    public function addLikeTo(HasLikesContract $model, $type, $userId)
+    public function addLikeTo(LikeableContract $likeable, $type, $userId)
     {
         $userId = $this->getLikerUserId($userId);
 
-        $like = $model->likesAndDislikes()->where([
+        $like = $likeable->likesAndDislikes()->where([
             'user_id' => $userId,
         ])->first();
 
         if (!$like) {
-            $model->likes()->create([
+            $likeable->likes()->create([
                 'user_id' => $userId,
                 'type_id' => $this->getLikeTypeId($type),
             ]);
@@ -61,7 +62,7 @@ class LikeableService implements LikeableServiceContract
 
         $like->delete();
 
-        $model->likes()->create([
+        $likeable->likes()->create([
             'user_id' => $userId,
             'type_id' => $this->getLikeTypeId($type),
         ]);
@@ -70,19 +71,17 @@ class LikeableService implements LikeableServiceContract
     /**
      * Remove a like to likeable model by user.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @param string $type
      * @param int|null $userId
      * @return void
      *
      * @throws \Cog\Likeable\Exceptions\LikerNotDefinedException
      */
-    public function removeLikeFrom(HasLikesContract $model, $type, $userId)
+    public function removeLikeFrom(LikeableContract $likeable, $type, $userId)
     {
-        $userId = $this->getLikerUserId($userId);
-
-        $like = $model->likesAndDislikes()->where([
-            'user_id' => $userId,
+        $like = $likeable->likesAndDislikes()->where([
+            'user_id' => $this->getLikerUserId($userId),
             'type_id' => $this->getLikeTypeId($type),
         ])->first();
 
@@ -96,38 +95,38 @@ class LikeableService implements LikeableServiceContract
     /**
      * Toggle like for model by the given user.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @param string $type
      * @param string $userId
      * @return void
      *
      * @throws \Cog\Likeable\Exceptions\LikerNotDefinedException
      */
-    public function toggleLikeOf(HasLikesContract $model, $type, $userId)
+    public function toggleLikeOf(LikeableContract $likeable, $type, $userId)
     {
         $userId = $this->getLikerUserId($userId);
 
-        $like = $model->likesAndDislikes()->where([
+        $like = $likeable->likesAndDislikes()->where([
             'user_id' => $userId,
             'type_id' => $this->getLikeTypeId($type),
         ])->exists();
 
         if ($like) {
-            $this->removeLikeFrom($model, $type, $userId);
+            $this->removeLikeFrom($likeable, $type, $userId);
         } else {
-            $this->addLikeTo($model, $type, $userId);
+            $this->addLikeTo($likeable, $type, $userId);
         }
     }
 
     /**
      * Has the user already liked likeable model.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @param string $type
      * @param int|null $userId
      * @return bool
      */
-    public function isLiked(HasLikesContract $model, $type, $userId)
+    public function isLiked(LikeableContract $likeable, $type, $userId)
     {
         if (is_null($userId)) {
             $userId = $this->loggedInUserId();
@@ -137,7 +136,7 @@ class LikeableService implements LikeableServiceContract
             return false;
         }
 
-        return $model->likesAndDislikes()->where([
+        return $likeable->likesAndDislikes()->where([
             'user_id' => $userId,
             'type_id' => $this->getLikeTypeId($type),
         ])->exists();
@@ -146,12 +145,12 @@ class LikeableService implements LikeableServiceContract
     /**
      * Decrement the total like count stored in the counter.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @return void
      */
-    public function decrementLikesCount(HasLikesContract $model)
+    public function decrementLikesCount(LikeableContract $likeable)
     {
-        $counter = $model->likesCounter()->first();
+        $counter = $likeable->likesCounter()->first();
 
         if (!$counter) {
             return;
@@ -163,15 +162,15 @@ class LikeableService implements LikeableServiceContract
     /**
      * Increment the total like count stored in the counter.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @return void
      */
-    public function incrementLikesCount(HasLikesContract $model)
+    public function incrementLikesCount(LikeableContract $likeable)
     {
-        $counter = $model->likesCounter()->first();
+        $counter = $likeable->likesCounter()->first();
 
         if (!$counter) {
-            $counter = $model->likesCounter()->create([
+            $counter = $likeable->likesCounter()->create([
                 'count' => 0,
                 'type_id' => LikeType::LIKE,
             ]);
@@ -183,12 +182,12 @@ class LikeableService implements LikeableServiceContract
     /**
      * Decrement the total dislike count stored in the counter.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @return void
      */
-    public function decrementDislikesCount(HasLikesContract $model)
+    public function decrementDislikesCount(LikeableContract $likeable)
     {
-        $counter = $model->dislikesCounter()->first();
+        $counter = $likeable->dislikesCounter()->first();
 
         if (!$counter) {
             return;
@@ -200,15 +199,15 @@ class LikeableService implements LikeableServiceContract
     /**
      * Increment the total dislike count stored in the counter.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @return void
      */
-    public function incrementDislikesCount(HasLikesContract $model)
+    public function incrementDislikesCount(LikeableContract $likeable)
     {
-        $counter = $model->dislikesCounter()->first();
+        $counter = $likeable->dislikesCounter()->first();
 
         if (!$counter) {
-            $counter = $model->dislikesCounter()->create([
+            $counter = $likeable->dislikesCounter()->create([
                 'count' => 0,
                 'type_id' => LikeType::DISLIKE,
             ]);
@@ -227,10 +226,12 @@ class LikeableService implements LikeableServiceContract
     public function removeLikeCountersOfType($likeableType, $type = null)
     {
         if (class_exists($likeableType)) {
-            $model = new $likeableType;
-            $likeableType = $model->getMorphClass();
+            /** @var \Cog\Likeable\Contracts\Likeable $likeable */
+            $likeable = new $likeableType;
+            $likeableType = $likeable->getMorphClass();
         }
 
+        /** @var \Illuminate\Database\Eloquent\Builder $counters */
         $counters = app(LikeCounterContract::class)->where('likeable_type', $likeableType);
         if (!is_null($type)) {
             $counters->where('type_id', $this->getLikeTypeId($type));
@@ -241,23 +242,53 @@ class LikeableService implements LikeableServiceContract
     /**
      * Remove all likes from likeable model.
      *
-     * @param \Cog\Likeable\Contracts\HasLikes $model
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
      * @param string $type
      * @return void
      */
-    public function removeModelLikes(HasLikesContract $model, $type)
+    public function removeModelLikes(LikeableContract $likeable, $type)
     {
         app(LikeContract::class)->where([
-            'likeable_id' => $model->getKey(),
-            'likeable_type' => $model->getMorphClass(),
+            'likeable_id' => $likeable->getKey(),
+            'likeable_type' => $likeable->getMorphClass(),
             'type_id' => $this->getLikeTypeId($type),
         ])->delete();
 
         app(LikeCounterContract::class)->where([
-            'likeable_id' => $model->getKey(),
-            'likeable_type' => $model->getMorphClass(),
+            'likeable_id' => $likeable->getKey(),
+            'likeable_type' => $likeable->getMorphClass(),
             'type_id' => $this->getLikeTypeId($type),
         ])->delete();
+    }
+
+    /**
+     * Get collection of users who liked entity.
+     *
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
+     * @return \Illuminate\Support\Collection
+     */
+    public function collectLikersOf(LikeableContract $likeable)
+    {
+        $userModel = $this->resolveUserModel();
+
+        $likersIds = $likeable->likes->pluck('user_id');
+
+        return $userModel::whereKey($likersIds)->get();
+    }
+
+    /**
+     * Get collection of users who disliked entity.
+     *
+     * @param \Cog\Likeable\Contracts\Likeable $likeable
+     * @return \Illuminate\Support\Collection
+     */
+    public function collectDislikersOf(LikeableContract $likeable)
+    {
+        $userModel = $this->resolveUserModel();
+
+        $likersIds = $likeable->dislikes->pluck('user_id');
+
+        return $userModel::whereKey($likersIds)->get();
     }
 
     /**
@@ -274,11 +305,9 @@ class LikeableService implements LikeableServiceContract
     {
         $userId = $this->getLikerUserId($userId);
 
-        return $query->whereHas('likesAndDislikes', function ($q) use ($type, $userId) {
-            $q->where([
-                'user_id' => $userId,
-                'type_id' => $this->getLikeTypeId($type),
-            ]);
+        return $query->whereHas('likesAndDislikes', function (Builder $innerQuery) use ($type, $userId) {
+            $innerQuery->where('user_id', $userId);
+            $innerQuery->where('type_id', $this->getLikeTypeId($type));
         });
     }
 
@@ -292,13 +321,14 @@ class LikeableService implements LikeableServiceContract
      */
     public function scopeOrderByLikesCount(Builder $query, $likeType, $direction = 'desc')
     {
-        $model = $query->getModel();
+        $likeable = $query->getModel();
 
         return $query
-            ->select($model->getTable() . '.*', 'like_counter.count')
-            ->leftJoin('like_counter', function ($join) use ($model, $likeType) {
-                $join->on('like_counter.likeable_id', '=', "{$model->getTable()}.{$model->getKeyName()}")
-                    ->where('like_counter.likeable_type', '=', $model->getMorphClass())
+            ->select($likeable->getTable() . '.*', 'like_counter.count')
+            ->leftJoin('like_counter', function (JoinClause $join) use ($likeable, $likeType) {
+                $join
+                    ->on('like_counter.likeable_id', '=', "{$likeable->getTable()}.{$likeable->getKeyName()}")
+                    ->where('like_counter.likeable_type', '=', $likeable->getMorphClass())
                     ->where('like_counter.type_id', '=', $this->getLikeTypeId($likeType));
             })
             ->orderBy('like_counter.count', $direction);
@@ -313,7 +343,8 @@ class LikeableService implements LikeableServiceContract
      */
     public function fetchLikesCounters($likeableType, $likeType)
     {
-        $likesCount = app(LikeContract::class)->query()
+        /** @var \Illuminate\Database\Eloquent\Builder $likesCount */
+        $likesCount = app(LikeContract::class)
             ->select([
                 DB::raw('COUNT(*) AS count'),
                 'likeable_type',
@@ -328,9 +359,7 @@ class LikeableService implements LikeableServiceContract
 
         $likesCount->groupBy('likeable_id', 'type_id');
 
-        $counters = $likesCount->get()->toArray();
-
-        return $counters;
+        return $likesCount->get()->toArray();
     }
 
     /**
@@ -367,6 +396,7 @@ class LikeableService implements LikeableServiceContract
     /**
      * Get like type id from name.
      *
+     * @todo move to Enum class
      * @param string $type
      * @return int
      *
@@ -380,5 +410,15 @@ class LikeableService implements LikeableServiceContract
         }
 
         return constant("\\Cog\\Likeable\\Enums\\LikeType::{$type}");
+    }
+
+    /**
+     * Retrieve User's model class name.
+     *
+     * @return \Illuminate\Contracts\Auth\Authenticatable
+     */
+    private function resolveUserModel()
+    {
+        return config('auth.providers.users.model');
     }
 }