+++ tests/Unit/Traits/LikeableTest.php
@@ -18,11 +18,11 @@ use Cog\Likeable\Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
 /**
- * Class HasLikesTest.
+ * Class LikeableTest.
  *
  * @package Cog\Likeable\Tests\Unit\Traits
  */
-class HasLikesTest extends TestCase
+class LikeableTest extends TestCase
 {
     use DatabaseTransactions;
 
@@ -568,5 +568,39 @@ class HasLikesTest extends TestCase
             $entityC->getKey() => '5',
             $entityD->getKey() => '10',
         ], $sortedEntities->pluck('count', 'id')->toArray());
+    }
+
+    /** @test */
+    public function it_can_collect_likers()
+    {
+        $entity = factory(Entity::class)->create();
+        $user1 = factory(User::class)->create();
+        $user2 = factory(User::class)->create();
+        $user3 = factory(User::class)->create();
+        $entity->like($user1->id);
+        $entity->dislike($user2->id);
+        $entity->like($user3->id);
+
+        $likers = $entity->collectLikers();
+
+        $this->assertCount(2, $likers);
+        $this->assertEquals([$user1->id, $user3->id], $likers->pluck('id')->toArray());
+    }
+
+    /** @test */
+    public function it_can_collect_dislikers()
+    {
+        $entity = factory(Entity::class)->create();
+        $user1 = factory(User::class)->create();
+        $user2 = factory(User::class)->create();
+        $user3 = factory(User::class)->create();
+        $entity->dislike($user1->id);
+        $entity->like($user2->id);
+        $entity->dislike($user3->id);
+
+        $dislikers = $entity->collectDislikers();
+
+        $this->assertCount(2, $dislikers);
+        $this->assertEquals([$user1->id, $user3->id], $dislikers->pluck('id')->toArray());
     }
 }